Feature/show participant submissions with results and allow to delete them (#746)
This change makes the participations details page more useful for instructors. They can see all submissions from a particular student together with the results (if available) and can potentially delete the submission.

This is helpful in case there are duplicated submissions (for whatever reason), or if the submission should be reset for whatever reason. It also helps the instructor to check if the submission of a student is actually in the system.

This page could be extended in the future, e.g. with a link to the actual submission and to the assessment (in case of modeling and text exercises).

And the page could also be made available to tutors without delete rights.